requirements.txt
appdirs==1.4.4
astroid==3.0.0
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
exceptiongroup==1.1.3
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
lazy-object-proxy==1.9.0
limits==3.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
ordered-set==4.1.0
packaging==23.2
pkg_resources==0.0.0
platformdirs==3.11.0
pluggy==1.3.0
psycopg2-binary==2.9.8
py==1.11.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pylint==3.0.0
pyparsing==3.1.1
pytest==7.4.2
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rich==13.6.0
six==1.16.0
SQLAlchemy==2.0.21
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
urllib3==2.0.6
watchdog==3.0.0
Werkzeug==3.0.0
wrapt==1.15.0
zipp==3.17.0
